pyproject.toml
[project]
name = "prompts-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.1",
    "fastapi>=0.129.0",
    "openai>=2.21.0",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.3",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
[tool.uv]
package = true

[project.scripts]
start = "uvicorn:main"
